refactor(type): 카드요청시 카드정보 타입, 카드요청 응답 에러타입 폴더 이동
