+++ src/feature/CardRegisterPage/types/card.ts
@@ -1,0 +1,8 @@
+import type { IssuerCodeType } from "../../../shared/types/Issuer";
+
+export type PostCardRequestBody = {
+  number: string;
+  expirationDate: string;
+  cvc: string;
+  issuerCode: IssuerCodeType;
+};
