+++ src/shared/types/api.ts
@@ -1,0 +1,4 @@
+export type RequestErrorInformation = {
+  code: string;
+  message: string;
+};
